Pass dump filename in as argument to s3 url script
Previously the filename was being set to a timestamp. The script is much
more useful if you can pass the filename in as an argument.

It allows for the stage to be prepended to the timestamp in the
Makefile, so we know which db a dump may have come from.