Add script to check the db dump is decryptable
There's a risk that if the secret key we use to decrypt the dumps gets
rotated, and the public key were encrpyting with doesn't, then we'll be
creating dumps we can never use.

This script will use the secret key in the creds repo to check that the
dumps can be decrypted. Using the `--list-packets` command with GPG
allows us to check that we can decrypt without actually having to
decrypt the file.