Pass recipient of encryption in as env variable
When encrypting asymmetrically with GPG you need to specify who the
recipient is. The encryption is then performed with their public key so
only they are able to decrypt it. When we're encrypting the db the
recipient is actually us, so we specify our own public key.

Setting the recipient to an env variable makes the script more usable.