Add make rule for creating db-cleanup service
The cleaning of the database dump will happen in a new, separate
PaaS space to production to help prevent any accidents on prod. The new
space is called `db-cleanup`.

The first thing the rule does is connect to preview and fetch the type
of postgres plan being used. The plan type is the same across preview,
staging and production.

We then switch to db-cleanup and create the postgres service we're going
to use, with the plan type we already fetched from preview.