Add make rule to deploy the db-cleanup app
First the rule checks to make sure that the currently targeted PaaS
space is db-cleanup. If its not we error out.

The lightweight db-cleanup app is then pushed to the space. It has no
route and no healthcheck as it's just a worker app that's allowing us to
tunnel through to the database service.

Once the app is pushed it's bound to the new db-cleanup service and then
restaged to make sure the new env variables describing its service are
available to it.

As this is such a simple app there's no need to have a manifest for it.