Add script to import and clean latest db dump
The script fill first query S3 for the latest production db dump and
download it. This will need to be performed by an entity with the
correct permissions on the bucket - Jenkins.

Then the secret gpg key required to decrypt the database dump is
imported.

The credentials and address of the db service are then fetched from
cloud foundry api so we can connect to the service.

Then a tunnel is created between local port 66306 and port 5432 on the
database host. This process is then kept open but sent to the background
and the PID stored so we can kill it later.

The sleep command is necessary to prevent failure due to a race
condition with the tunnel.

The downloaded dump is then decrypted and unzipped and streamed to the
db service via psql.

An sql script is then executed on the db service to clean it of
sensitive data. The --variable flag allows us
to pass in a password hashed with bcrypt. The hashing is done by a
separate python script.

Finally we kill the ssh tunnel to the host.