Mitigate race condition when creating cleanup service
When switching targets the subshell will return before the space has
actually changed. This resulted in the DB_PLAN variable not being
set properly on the following line. Adding a sleep here mitigates that.

Also adding a check to make sure that the db-cleanup space is active
before creating the service.