Upload cleaned dump to target stage
We connect to the PaaS space specificed by the target stage, and upload
the cleaned db dump to it's database. We make the connection is the same
way as done before.

There's a check on the target stage to ensure that it can only be
applied to preview or staging.

Alternatively, just uploading to google drive is possible.