Make sure declaration have `organisationSize`
Organisation size is now being replayed to users in a feature to show
who has applied for what opportunities. It gets the size from the
declaration. Without it, the functional tests will fail on preview and
staging. We add it back here.

The MOD command takes a modulus 4 of the organisation name to select a
size from the array. It's good as it will always choose the same size if
their name remains the same. Credit to Ben Vandersteen for that one.