Create postgres app instead of using service
Spinning up a service takes ages. It also means that functions are
created by default (`reassign_owned`) which cause us problems. It also
means that any commands we run on the database are not done as a super
user.

We need to remove the `reassign_owned` function to make cleaned db dumps
play nicely with local dev environments.

We can just push a postgres image and work directly on that which solves
all of the above problems.

We currently can't create tcp apps in PaaS which is why we need to
tunnel via ssh to the app instead of being able to just use psql
directly.