Sanitise rather than delete framework agreements
We can sanitise the personal details (signer name and role, and the S3
paths to actual signed/countersigned documents) but preserve the fact that
the agreement has been returned.

This means that when logging in as a supplier in Preview or Staging
environments the framework dashboard will be accessible without having
to go through the whole "return your framework agreement" flow.

This will be useful for developers working on the framework dashboard
who do not want to repeatedly return fake framework agreements for
suppliers in order to test their new features.