Explicitly list full routes in PaaS vars
Changes the base manifest template to include application routes
as-is (only formatting them to add current environment) instead of
constructing the route from domain / subdomain and paths.

This allows us to specify multiple subdomains / domains per app,
and specify manifests for apps that don't need a route.

To avoid duplicating the application paths between environments we
format routes with the current environment, so they can be specified
once in the common vars file.