Use 'max' instead of 'sum' in dashboards
Normal graphite uses a 'last-write-wins' setup so that you can overwrite
datapoints with the same timestamp. Hosted Graphite doesn't. This is so
that you don't have to preaggregate values to match the resolution of
the dashboards.

When sending metrics, out app collects that last 10 minutes metrics and
sends them. This means you get duplicates. And Hosted Graphite merges
them all for you. Using max prevents artificially inflated values.

Also, move 500s count to the bottom row so it's wider. It seems more
important.