Don't set default values. Yet.
Setting a default value of zero means that we won't get patch data.
However, we're currently sending samplecount to Hosted Graphite from
Cloudwatch. This means that if a default value of zero is recorded, it
counts as a 'hit' on the samplecount statistic.

If we set the default value from the beginning, the data would be skewed
in the HG dashboards until they were changed to use the 'sum' instead of
'samplecount'.

It makes more sense to get the dashboards working with `sum` and no
default value (which should be identical to `samplecount`), and then
turn on the default value which will start filling in the blanks.