Set log-metric default values to 0
Now that we're using `sum` rather than `samplecount` in our Hosted
Graphite setup, we can start sending a default value for log metrics.

The default value will be set for a metric if the metric filter pattern
can't be found in any of the logs. As a note, if there are NO logs for
the period, then no metric will be recorded.