Add script to create missing log alerts in Hosted Graphite
We want to know if our logging infrastructure breaks. We can do
this by monitoring AWS `incomingLogEvents` metrics. If no logs
are being added to AWS Cloudwatch log groups for a period of time
then we can trigger an alert.

Note, we looked into setting these up as AWS Cloudwatch alerts
but ended up not going down that path. This was because
- it would be another type of infrastructure to use
- slack alerts would be extra work to set up than emails which
differs from our current alerts
- terraform doesn't let you set up cloudwatch alerts that go to
an email address (email is an unsupported protocol as described
https://www.terraform.io/docs/providers/aws/r/sns_topic_subscription.html)
unless the email alert is set up manually without using terraform