Add Jenkins_2
This adds the Terraform code to spin up a new instance for a new Jenkins
box. This is being done to allow pen testers to test a Jenkins instance
that isn't our actual instance. It will be provisioned via Ansiible to
be a similar as we can get it.

This Terraform code uses the existing security group that is applied to
our current box, and the same instance profile.

There is an extra security group applied, which is to allow access to
the box from Cyberis's IP's.

This could lay the groundwork for a new, upgraded Jenkins box which
could be built in the future. Alternatively, this Terraform could be
scrapped and we could move Jenkins to the PaaS.