Bump provider version to 1.9.0
The main driver for this was a bug that has been fixed in 1.9.0.
vpc_security_group_ids were not being persisted in the statefile and so
we're being reapplied every run. More details are here:
https://github.com/terraform-providers/terraform-provider-aws/pull/2338

This release also deprecated the use of the `current` attribute in the
data.aws_region block. It defaults to current now instead, so it can be
removed.

I've applied this bump across all environments/accounts as it makes
sense to keep these in sync.