Force destroy IAM users if we delete them
We recently ran into a problem when trying to remove an IAM user.
It would complain that terraform couldn't remove the user because
they had existing login-profiles that needed to be deleted first.
This may be because login profiles were set up manually.

The solution to this is for our users to be marked as `force_destroy`
so that if we choose to delete them, terraform will also destroy
any of their related login profiles and keys.

https://www.terraform.io/docs/providers/aws/r/iam_user.html#force_destroy

https://github.com/hashicorp/terraform/issues/5908