Move maintenance mode variable
Rather than defining maintenance mode in the common vars, which will
make it affect all environments simultaneously, we should define it on a
per-environment basis so that we are able to toggle it for specific
environments without the risk of accidentally also turning it on for
production.

This will allow us to more safely have a job that can toggle maintenance
mode on/off.