Give Jenkins role permissions for prod submissions bucket
We'd like Jenkins to run the virus scanning job for uploaded files. The
Jenkins role needs the correct permissions to do this.

The other part of this adding permissions on the bucket to allow Jenkins
to list it. We don't currently manage those buckets via Terraform so
this has been done through the GUI via the S3 Access Control List.