Give Jenkins role permissions on comms buckets
We have a job on Jenkins that runs daily that uploads a csv of DOS
opportunity details to the preview and production communications
buckets.

Currently, the script that does this is having AWS credentials injected
into it as env variables. These creds are coming from the jenkins users
aws config file. That config file is a template populated by Ansible,
which gets the creds from the jenkins vars file in the credentials
repo.

The AWS creds in the credentials repo are for an AWS user we have set up
called `deploy`. This user is a hangover from when we used to use
ElasticBeanstalk, and actually has quite a lot or permissions to do
stuff. We should get rid of it, as this script appears to be the only
thing using it.

The script doesn't actually need env variables injecting in, as boto can
use the Jenkins instance profile. We just need to give the jenkins role
the correct permissions, which is what this is doing.

We need `GetObject`, as the file being uploaded is overwriting an already
existing file and it reads it's metadata via the `HeadObject` operation.
We need `PutObjectAcl` as the objects permissions are set to public.

The other part of this is updating the buckets permissons to allow
`PutObject` from the Jenkins role. Currently we don't manage these
buckets through Terraform, so I will do it by hand. But we should think
about bringing those buckets into the Terraform fold.