Scale supplier and api for DOS3 applications
This increases the supplier frontend and api instance counts to match
what we were running for G-Cloud 10 applications.

We currently have a 200GB memory limit on PaaS or which we're currently
using about 26GB. This bump will add at most an extra 64GB of memory use
(if we were to release both applications at the same time), so we should
still be well within limits.