Add metric filter for apiclient retries
Now that we're retrying requests by the apiclient if they fail for some
reason (most often a connection error), we're no longer getting the
alerts for the 500s that resulted. This is good, but we have no insight
into when we're having to retry stuff.

We can create a metric for the log messages for urllib3.util.retry which
will let us measure how often we need to retry stuff.