Put Jenkins3 behind ELB
This was done to get a free certificate, rather than having to get one
from letsencrypt. The main motivation was I kept bumping in to the
letsencrypt rate limit on issuing certificates of 20 per week.

This code sets up an CNAME record for ci3.market.... which points to the
DNS name of the elb. The elb terminates our tls and proxies the request
on to the Jenkins instance via http. This is all within our VPC and
subnet.

The ELB also passes tcp connections through port 22 to allow ssh access.

There are two security groups - one for the ELB with all our usual
whitelisted IP address as the only allowed ingress, and one for the
instance itself, which only allowed ingress from the ELB.