Add sns topic subscription
This adds a subscription to the file upload topic for an endpoint on the
antivirus api. The api will be using basic auth which is included in the
endpoint described here. The creds are in the creds repo.

The endpoint _has_ to be able to autoconfirm the subscription request
that is sent to it or this code will not apply.