Refactor Jenkins module
This refactors the Jenkins module to make it actually reusable.
Previously it defined resources within the module that would not be
duplicatable had an extra module been created.

The Jenkins instance profile has been moved out of the module and into
the main account. We only want one of these - quite a few things depend
on knowing the exact arn of the jenkins instance profile so it doesn't
make sense to create a new one.

The A record for Jenkins has been moved out of the route53 file in the
main account and into the module. This wasn't strictly neccessary, but
it keeps all the changeable definitions for the module in one place.

The Jenkins EC2 instance is now defined with variables that are passed
in so different modules can use different values for ami and instance
type. Useful if upgrading.

The wildcard certificate for Jenkins ELB has moved out of the module and
into the main account. It's a wildcard, so we can't/don't need to define
it more than once for different modules/subdomains.