Add ability to specify PaaS manifest template variables on command line
Adds an option flag to `generate-paas-manifest.py` that can take
variables in the form `option=value` that are passed into the template
context.

Additionally, a value of `env_var` has its value read from the
environment, which negates the need to expose possible secrets on the
command line.

We also add an optional `ARGS` envvar to the make command so these
options can be passed through.