383-recovery-alarms-on-new-aws-alerts
Add recovery emails for alarms in terraform #503

When an AWS alarm triggers a notification it's actually triggering a 'topic'.
A topic is just a list of emails that will all get the same notification.

In our case we want one email address to receive bad :( notifications and one email address to receive good :) email notifications.

To do this we have to create 2 'topics', a 'good' topic and a 'bad' topic, each with one email registered to it.

When a good notification is triggered it is sent to the 1 slack integration email subscribed to that 'topic'.
And the 1 slack integration email subscribed to our bad 'topic' gets the bad notifications.

This allows us to show them as green and red in Slack because you can only show icons for different users or emails, not based on the content of the message.

In this PR we create the new 'good' topic, give it the same policy as the bad one (only gds emails can subscribe, only actions in our AWS accounts can trigger emails etc), and then tell AWS to notify the new 'good' topic when an alarm changes to the OK state.