deploy-app: add network policy handling
using a new app var key `egress_to_applications` that can be used to define
which policies should be created. of course, an application being released
has to ensure new network policies are created both for any applications it
is wanting to access, but also any applications that are wanting to access
*it*. cue some yq magick to process the yml to automatically determine
these "ingress" relationships.

the order of operations performed here is designed to make it impossible
for an app to get to the point of being started with any required network
policies not being in place, even if the "other" app (the other "party" in
the policy) is simultaneously going through the release process and having
its names & routesthe order of operations performed here is designed to make it impossible
for an app to get to the point of being started with any required network
policies not being in place, even if the "other" app (the other "party" in
the policy) is simultaneously going through the release process and having
its names & routes similarly juggled around.
 similarly juggled around.