Change deploy-app to push new app without route
Our deploy-app recipe will create a `-release` app with a route and only
removes the route after it has been started. If the app start fails this
means there can be a route pointing to a dead app.

This commit changes the deploy-app recipe so that the release app will
be created without a route, and the route only added once it has
successfully started.

The route mapping is done by a new script, `scripts/map-route.sh`, which
parses the desired route(s) from the app manifest and adds them with `cf
map-route` without restarting the app.