Add digitalmarketplace-gov-uk to CodeCommit
We want to be able to backup the digitalmarketplace-govuk-frontend
repository to AWS CodeCommit, so we need to add it to the list of repos
here.

We add it to the end because otherwise Terraform will try to recreate
all the resources for repositories that follow in the list if we insert
in the correct alphabetical location.