Double antivirus-api instances for load tests
The load test we ran today was broadly successful. We found though that antivirus-api seemed to be struggling a bit. This isn't actually a problem - the overnight job will pick up anything that it missed during the day.

Double the number of instances ahead of our test tomorrow to see if that helps things.