Document need to update production rate limit
nginx does rate limiting on a per-instance basis. So if you increase the number of instances, you need to decrease the rte limit so that the overall rate limit stays the same. Add a comment reminding people to do this.

The same thing could also apply for other environments, but we only really care about production.