Update `pg_dump` in db-backup container
We are now using PostgreSQL 10.13 in production. `pg_dump` will refuse
to connect to servers newer than itself[[1]], so we need to upgrade it.
To do this we upgrade the base image to one that uses Debian Buster
(stable), and then install `postgresql-client` from the Debian repos
(rather than the postgresql repos). This currently installs,
postgresql client 11, which will work with server 10.

[[1]]: https://www.postgresql.org/docs/11/app-pgdump.html#PG-DUMP-NOTES