Convert input files to tfvars format
Previously, we were using JSON files. However, as of terraform 0.12, terraform starts trying to parse them as JSON. This happens because the file descriptor that sops-wrapper produces as the filename doesn't end with '.json'.

Rather than try to wrangle that mess, the easiest solution is to switch to tfvars format. This has the added benefit of being slightly easier for humans to work with than JSON.

Depends on https://github.com/alphagov/digitalmarketplace-credentials/pull/368