Refactor Jenkins Route53 DNS validation
In v3 and up of the terraform AWS provider, the `domain_validation_options`
property of an `aws_acm_certificate` is now an optional list of
values to allow for manual email verification. This means we can
no longer access values by index position.

Refactor to instead use `for_each` to iterate through all potential
values in the list.