Upgrade backups account to AWS provider v3.34.0
We only need to define the provider version in the places we will
be running Terraform from. Use the new `required_providers` syntax
so this change is compatible with both Terraform v0.12 and v0.13.

This account uses two AWS regions. Set the 'main' provider as
`eu-west-1` using the new syntax and create an alias `aws.london`
that uses the same provider version but in `eu-west-2` as the
backup location.

S3 buckets no longer need a `region` property - they inherit the
region from the provider.