Upgrade production environment to AWS provider v3.34.0
We only need to define the provider version in the places we will
be running Terraform from. Use the new `required_providers` syntax
so this change is compatible with both Terraform v0.12 and v0.13.

Remove the region tags from the cross-region backup buckets. The
region is now computed from the one set for the provider.