Set AWS_DEFAULT_REGION for validation
This is a workaround for a limitation in Terraform's language validation.
Each module is validated separately, and since we have moved the
provider definitions out of modules Terraform can't see the region
defined in the providers that the module inherits. Setting a
default region via environment variable allows the validation to
pass.

See https://github.com/hashicorp/terraform/issues/21408 for more details.