Fix CloudWatch log group ARN
In #874 we upgraded the Terraform AWS provider to 3.34.0. In the
PR description, I noted that this upgrade made changes to the
CloudWatch log group ARN, but that I didn't think it mattered.

It turns out this isn't true - AWS rejects the ARN without a
trailing `:*` but this only raises an error when we try to apply
the changes.

From the provider upgrade docs[[1]]:

> Previously, the resource returned the Amazon Resource Name (ARN) directly from the API, which included a :* suffix to denote all CloudWatch Log Streams under the CloudWatch Log Group. Most other AWS resources that return ARNs and many other AWS services do not use the :* suffix. The suffix is now automatically removed.

The docs recommend using string interpolation to add the trailing
`:*` back in everywhere it was missing, so do that.

[1]: https://registry.terraform.io/providers/hashicorp/aws/latest/docs/guides/version-3-upgrade#resource-aws_cloudwatch_log_group