Add another router app to Staging and Preview
We need to decrease the production rate limit to cope with an over-eager scraper. See https://github.com/alphagov/digitalmarketplace-router/pull/109

Unfortunately, this lower rate limit is just too low for Preview and Staging. They typically see peaks of ~8 requests/s from our tests. The rate limit change reduces their rate limits from 10 requests/s to 6 requests/s.

So increase the number of router instances to 3. This should increase the rate limit for those environments to 9 requests/s, which should be sufficient for our tests.

This is intended to be a temporary stop-gap whilst we work out a better long-term approach to coping with the scraper's traffic.