Start using invoke for Python-based tasks
So that we can manage this repo in the same way we do all our others.

See https://github.com/alphagov/digitalmarketplace-scripts/pull/649 for an example of how we've done this for other repos. There's a lot of non-Python stuff in the Makefile, so I've left that as-is for the moment in the interests of time.

As part of this, I needed to switch us over to using pip-compile. This has no immediate effect, but might make future dependency management easier.