Load bins from venv at end of path
We added venv/bin to the path so that scripts can use tools we've added to the venv such as yq (426d3dc62376369a9d5fca1b58eb5ca64db1bf66). This works fine for those tools. However, it also means that we preferentially use the python and pip inside the venv. This causes problems when installing invoke and the developer tools, which expect to be installed outside the venv.

So load the venv bins into the path last. This should have no impact on the tools in the venv. It will mean that we will now default to the system python/pip. This is the correct behavioue because the bootstrap command is the only one that uses python/pip. So this change should fix the bootstrap command without breaking anything else.