Correct name of ACM validation DNS entry
Previously, the ACM DNS entry for assets was using the same name as for www. I think this was a copy-paste error, because there exists an assets_acm_name variable that was defined but not used.

I think this should put a stop to the current dueling between assets and www, where terraform flips back and forth because they both have the same name.